safari-534.53-branch/LayoutTests/editing/pasteboard/paste-blockquote-into-blockquote-3.html

<html>
<head>
    <style>
    blockquote {
        color: blue;
        border-left: 2px solid blue;
        margin: 0px;
        padding: 0 0 0 20px;
    }
    </style>
</head>
<body contenteditable>

<div>
    <div>
        <p>This tests pasting a blockquote into a blockquote. Nothing below should be double-blockquoted.</p>
        <blockquote type="cite">one<br>two<br>three<br id="endContent"><br><br></blockquote>
    </div>
    <br>
</div>

<script>
var sel = window.getSelection();
var endContent = document.getElementById("endContent");
sel.setPosition(endContent, 1);
document.execCommand("InsertHTML", false, "<blockquote type='cite'><div>two</div></blockquote>");
</script>

</body>
</html>
